refactoring for (hopefully) improved ease of use
